resources: support edge rule resources
Add support for managing Pull Zone Edge Rules via the Terraform provider.
Edge Rules are represented as a separate Terraform resource that are linked via
pull_zone_id attribute to a Pull Zone.

All edge rule fields are fully supported except for the enabled and description
fields.
The enabled field is currently are computed field, support for it will be added
in a follow-up.
The description field can currently not be set because it is used by terraform
provider to identify newly created edge rules. When edge rules are created via
the bunny.net API, the API does not return the definition or the GUID of the
edge rule. The Pull Zone must be retrieved afterwards and the new edge rule must
be found in the list to get it's GUID. To be able to find newly created edge
rules reliably, the provider stores in the Description field of an Edge Rule an
internally generated uuid and then searches for an edge rule with this uuid in
the retrieved pull zone.
This identification must only be done once after creation. Therefore the
description field could be made available to the user, in a follow up, by
changing it to the user defined value in a second step during creation.

When a trigger of an edge rule is updated, Terraform shows a diff that recreates
all triggers. This issue will be fixed in a follow-up.