scripts: add a goreleaser config + script to create releases
This commit introduces a goreleaser config, based on the config from the
terraform-provider-scaffold repository and shell script to create releases.

The shell script finalizes the CHANGELOG.md file, creates a gpg-signed git tag
and runs goreleaser.
Goreleaser creates release binaries, a signed checksum file and creates a draft
release on github with the release binaries attached.