makefile: set version to dev
When building releases the version is set via goreleaser.

When compiling the binary via the makefile set the version to the string "dev".
The version of local build binaries is not relevant, it often contains local
changes and therefore is not the same then a release binary with the same
version.