makefile: add a gen-dev-tftrc target
Add a target called gen-dev-tftrc to the Makefile.
When invoked it calls a bash-script which generates a terraform config file with
a dev_overrides statement that references the INSTALLDIR used in make install.

Via "make install gen-dev-tftrc", a local provider binary can now be installed
and a config be generated that can used to use the local binary with terraform.