resources: add pull-zone hostname resource
Add support for managing custom hostnames of pull zone via a new hostname
resources.

The bunny APi provides a seperate endpoint for adding hostnames to a pull zone.
The endpoint does not return the ID of the created hostname, therefore it is
retrieved by an additional call to the "Add Pull Zone Endpoint".
Hostnames have only a single setting (force_ssl), apart from the hostname
itself, that can be set. It can not be specified when creating the hostname.
It must be enabled/disabled via the seperate "Set Force SSL" endpoint.
The endpoint is not supported yet by the bunny-go api package.
It will be added in a follow-up PR. Until then the force_ssl field is read-only.

Related API documentation:
https://docs.bunny.net/reference/pullzonepublic_index2
https://docs.bunny.net/reference/pullzonepublic_addhostname
https://docs.bunny.net/reference/pullzonepublic_removehostname
https://docs.bunny.net/reference/pullzonepublic_setforcessl