hostname: support load_free_certificate
Add a new boolean load_free_certificate field.
If it is set to true, the provider calls the LoadFreeCertificate bunny API
endpoint to generate and load a free cert for the hostname.

Generating the free certificate requires that hostname has a CNAME record
pointing to a bunny.net domain. Because it can take some time until dns changes
are distrubted, loading the free cert is retried, when an error message is
returned that indicates the dns record is missing.
It uses Terraforms default retry timeout.

The bunny api does not provide any interface to get the status of
LoadFreeCertificate or to disable the flag after enabling it once.
Therefore the hostname resource is recreated when it's value changes.