tests: distinguish between nil and unset values when comparing states
When the state at the provider and the local state was compared, nil and default
values were evaluated as equal.
This can hide bugs. When e.g. a field is expected to be "false" but is nil the
comparision would have been evaluataed to equal.

Distinguish unset and default values.