pullzone: group headers related settings in a headers groups
To improve the structure of the pullzone resource, all header relating settings
are moved to a block called "headers".

The Terraform sdk has no type for structs/objects. Other providers commonly
represent them as a TypeList with with a single map[string]interface{} element.
To manage those kind of values a new struct called "structure" is introduced.