pullzone: change enable_access_control_origin_header type to string
The type of the enable_access_control_origin_header field is changed from a
string set to string.

The change is done because the terraform sdk does not support to set default
values for string sets.
Bunny uses a default value for the attribute.
If we can not set a default value for it, Terraform calculate a diff and wants
to set the field to nil/empty when it is not explicitly defined in the headers
block.

A better behaviour is to use the default from bunny.net if it is not explicitly
defined. This is possible when using a string type.