hostname: fix: plan not empty after hostname creation
After creating a new hostname a following  "terraform plan" command would show a
non-empty diff.
The values of the computed fields of a hostname were not retrieved and set in
the ResourceData after the creation.

Retrieve full hostname after creation and set the ResourceData to the retrieved
values from the API.