edgerule: fix: moving edge-rules to another pull-zone fails
When the pull_zone_id field of an edge rule was changed, e.g. to move an
edge rule to another pull zone, the operation failed with the error:
	updating edge rule failed: http-request to
	https://api.bunny.net/pullzone/XXXXXY/edgerules/addOrUpdate failed: Not
	Found (404)

Following "terraform apply" runs failed with the error:
	"pull zone has no edge rules"

Moving an edge-rule is not supported, it must be recreated for another
pull-zone.
The resource definition lacked to enable the ForceNew behaviour for the
pull_zone_id field, which caused the issue.