hostname: fix: local state incorrect when uploading certificate fails
When a new hostname with a certificate was created and uploading the certificate
failed, the local state was invalid.
The hostname was created successfully at bunny.net but it was missing from
the state. This resulted in hostname already exist errors on following
applies for the hostname.

Change the provider to not return an error immediately when uploading a
certificate fails. Following independent API operations for the same hostname
are still run, the state is set according and at the end the error about the
failed certificate upload is returned.
This results in a correct state.
The hostname resource will still be tainted by Terraform. This has the unwanted
effect that also when the broken certificate section is removed and therefore
the state and definition are the same the hostname resource is unnecessary
recreated. This can prevented manually by the user by untainting the resource.
There seems to be no good solution for this kind of issue[1].
Returning a warning instead of an error will result in a misleading success and
the warning would be easy to miss.

- [1]: https://github.com/hashicorp/terraform-plugin-sdk/issues/330.