tests: fix: AccHostname_StateIsValidWhenCertUploadFails test fails
The TestAccHostname_StateIsValidWhenCertUploadFails started to fail because the
bunny.net API is not returning an error anymore when invalid data is posted as
certificate.

To generate an API error we post 800kiB of zeroes instead, which causes a 500
error.
This is the lowest value that I've tested that generates an error, 700kiB does
not.