pullzone: make origin_url & storage_zone_id mutually exclusive
The docs say that on creation of a PZ an OriginURL is required and a StorageZoneId is optional. 
However, the reality is that ONE of the TWO are required. You must either specify a OriginURL OR a StorageZoneID. Furthermore, passing BOTH results in OriginURL being entirely ignored (StorageZoneID is preferred).
(https://github.com/simplesurance/bunny-go/pull/19#issuecomment-1146304345)