tests: ignore new unsupported pull-zone fields
Add the new pull-zone fields that are supported in the updated bunny-go version
to the diffIgnoredFields of the pull zone testcases.
They are not supported by the terraform provider yet.
Adding them to ignore list prevents that tests fail because those fields are
missing the terraform provider representation of a pull zone.