simplify storageZoneFromResource
storageZoneFromResource was only setting fields in
*bunny.StorageZoneUpdateOptions that changed in the resource.
Change it to set always all fields to the current values.
This is consistent with how other *FromResource functions are behaving.

The update message sent to the bunny API will contain all settings also the
current ones instead of only the changed ones.
Depending on how the bunny.net endpoint handles missing fields in JSON payloads
for update messages this could be better or worse. :-)
If it always interprets missing fields as no change is done, it causes only
unnecessary network traffic.
If it interprets missing fields as the setting should be unset/disabled, it's
better to include the whole configuration.

The bunny.net API distinguishes between missing and empty string values
for the OriginURL and Custom404FilePath fields. When submitting empty strings
for those it resulted in an error, that those fields set to invalid
values.

To handle it the helper function getOkStrPtr() is introduced. It returns nil if
the value is unset instead of an empty string.

Replacing all get<TYPE>Ptr() calls with getOk<TYPE>Ptr() might make sense and
should be investigated in a follow-up.