storagezone: use d.Partial in Update instead of revertUpdateValues
That after a failed update changes to the resource are still applied is a know
bug (https://github.com/hashicorp/terraform-plugin-sdk/issues/476).
It affects attributes that are not retrieved via the Get function and set only
by the provider.
custom_404_file_path is such a field.
As workaround d.Partial() can be called.
This is more simple then the revertUpdateValues() implementation.